requirements.txt
# Function dependencies, for example:
# package>=version
google-api-core==1.29.0
google-auth==1.30.1
google-cloud-core==1.6.0
google-cloud-storage==1.38.0
google-crc32c==1.1.2
google-resumable-media==1.3.0
googleapis-common-protos==1.53.0
python-dotenv==0.17.1
lxml==4.6.3
psycopg2==2.8.6
